backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.8.3
httpx==0.28.1
python-multipart==0.0.20
